refactor readBypassSecret into root vitest.config
no need for this to be in two places
