+++ packages/stream-metadata/vitest.config.ts
@@ -1,21 +1,5 @@
 import { defineConfig, mergeConfig } from 'vitest/config'
-import { readFileSync } from 'fs'
-import { resolve } from 'path'
-import { rootConfig } from '../../vitest.config.mjs'
-
-function readBypassSecret(): string | undefined {
-	try {
-		const riverEnv = process.env.RIVER_ENV || 'local_multi'
-		const runEnv = riverEnv.replace('local_', '')
-		const contractsPath = resolve(__dirname, `../../core/run_files/${runEnv}/contracts.env`)
-		const content = readFileSync(contractsPath, 'utf8')
-		const line = content
-			.split(/\r?\n/)
-			.find((l) => l.startsWith('RIVER_TESTENTITLEMENTSBYPASSSECRET='))
-		if (line) return line.split('=', 2)[1]
-	} catch {}
-	return undefined
-}
+import { rootConfig, readBypassSecret } from '../../vitest.config.mjs'
 
 export default mergeConfig(
 	rootConfig,
@@ -25,6 +9,7 @@ export default mergeConfig(
 			setupFiles: './vitest.setup.ts',
 			env: {
 				RIVER_ENV: process.env.RIVER_ENV || 'local_multi',
+				// skip entitlements for these tests
 				RIVER_TEST_ENT_BYPASS_SECRET: readBypassSecret() ?? '',
 			},
 		},
