+++ vitest.config.mts
@@ -2,6 +2,22 @@ import path from 'path'
 import os from 'os'
 import { defineConfig } from 'vitest/config'
 import wasm from 'vite-plugin-wasm'
+import { readFileSync } from 'fs'
+import { resolve } from 'path'
+
+export function readBypassSecret(): string | undefined {
+    try {
+        const riverEnv = process.env.RIVER_ENV || 'local_multi'
+        const runEnv = riverEnv.replace('local_', '')
+        const contractsPath = resolve(__dirname, `./core/run_files/${runEnv}/contracts.env`)
+        const content = readFileSync(contractsPath, 'utf8')
+        const line = content
+            .split(/\r?\n/)
+            .find((l) => l.startsWith('RIVER_TESTENTITLEMENTSBYPASSSECRET='))
+        if (line) return line.split('=', 2)[1]
+    } catch {}
+    return undefined
+}
 
 export const rootConfig = defineConfig({
     test: {
